+++ src/api-cuda.h
@@ -59,9 +59,9 @@ typedef struct FTIT_data_prefetch {
     // Pointer pointing to data to be fetched;
     unsigned char *dptr;
     // total number of bytes to be streamed;
-    size_t totalBytesToFetch;
+    int64_t totalBytesToFetch;
     // Bytes I will fetch each time;
-    size_t fetchSize;
+    int64_t fetchSize;
     // Check Whether Bytes are stored in device
     int isDevice;
     // Points to the current stream
@@ -77,25 +77,25 @@ typedef struct FTIT_data_prefetch {
 } FTIT_data_prefetch;
 
 
-typedef int (*FTIT_fwritefunc)(void *src, size_t size, void *opaque);
+typedef int (*FTIT_fwritefunc)(void *src, int64_t size, void *opaque);
 
 int FTI_get_pointer_info(const void *ptr, FTIT_ptrinfo *ptrInfo);
-int FTI_copy_from_device(void *dst, const void *src, size_t count,
+int FTI_copy_from_device(void *dst, const void *src, int64_t count,
  FTIT_execution *exec);
-int FTI_copy_to_device(void *dst, const void *src, size_t count,
+int FTI_copy_to_device(void *dst, const void *src, int64_t count,
  FTIT_execution *exec);
 int FTI_TransferDeviceMemToFileAsync(FTIT_dataset *data,
  FTIT_fwritefunc fwritefunc, void *opaque);
 int FTI_InitPrefetcher(FTIT_data_prefetch *dfls);
-int FTI_getPrefetchedData(FTIT_data_prefetch *dfls, size_t *size,
+int FTI_getPrefetchedData(FTIT_data_prefetch *dfls, int64_t *size,
  unsigned char **fetchedData);
 int FTI_destroyPrefetcher(FTIT_data_prefetch *dfls);
 int FTI_DestroyDevices();
 int FTI_InitDevices(int HostBuffSize);
 int FTI_TransferFileToDeviceAsync(FILE *fd, void *dptr, int numBytes);
 BYTE *FTI_getHostBuffer(int id);
-size_t FTI_getHostBuffSize();
-int FTI_copy_to_device_async(void *dst, const void *src, size_t count);
+int64_t FTI_getHostBuffSize();
+int FTI_copy_to_device_async(void *dst, const void *src, int64_t count);
 int FTI_device_sync();
 
 #endif  // FTI_SRC_API_CUDA_H_
